Upload local consents during signIn
* Verified in manual testing
* Unit tests broken
